+++ hercules-ci-agent/cbits/nix-2.4/hercules-store.hh
@@ -7,7 +7,6 @@
 #include <nix/common-eval-args.hh>
 #include <nix/get-drvs.hh>
 #include <nix/derivations.hh>
-#include <nix/affinity.hh>
 #include <nix/globals.hh>
 #include "HsFFI.h"
 
@@ -57,15 +56,42 @@ public:
   virtual void addToStore(const ValidPathInfo & info, Source & narSource,
       RepairFlag repair = NoRepair, CheckSigsFlag checkSigs = CheckSigs) override;
 
-  virtual StorePath addToStore(const string & name, const Path & srcPath,
+  virtual StorePath addToStore(
+#if NIX_IS_AT_LEAST(2,7,0)
+      std::string_view name,
+#else
+      const std::string & name,
+#endif
+      const Path & srcPath,
       FileIngestionMethod method = FileIngestionMethod::Recursive, HashType hashAlgo = htSHA256,
-      PathFilter & filter = defaultPathFilter, RepairFlag repair = NoRepair) override;
+      PathFilter & filter = defaultPathFilter, RepairFlag repair = NoRepair
+#if NIX_IS_AT_LEAST(2,5,0)
+      , const StorePathSet & references = StorePathSet()
+#endif
+      ) override;
 
-  virtual StorePath addToStoreFromDump(Source & dump, const string & name,
-      FileIngestionMethod method = FileIngestionMethod::Recursive, HashType hashAlgo = htSHA256, RepairFlag repair = NoRepair) override;
+  virtual StorePath addToStoreFromDump(
+      Source & dump, 
+#if NIX_IS_AT_LEAST(2,7,0)
+      std::string_view name,
+#else
+      const std::string & name,
+#endif
+      FileIngestionMethod method = FileIngestionMethod::Recursive,
+      HashType hashAlgo = htSHA256,
+      RepairFlag repair = NoRepair
+#if NIX_IS_AT_LEAST(2,5,0)
+      , const StorePathSet & references = StorePathSet()
+#endif
+      ) override;
 
-  virtual StorePath addTextToStore(const string & name, const string & s,
-      const StorePathSet & references, RepairFlag repair = NoRepair) override;
+  virtual StorePath addTextToStore(
+#if NIX_IS_AT_LEAST(2,7,0)
+    std::string_view name, std::string_view s,
+#else
+    const std::string & name, const std::string & s,
+#endif
+    const StorePathSet & references, RepairFlag repair = NoRepair) override;
 
   virtual void narFromPath(const StorePath & path, Sink & sink) override;
 
@@ -81,13 +107,17 @@ public:
 
   virtual void addTempRoot(const StorePath & path) override;
 
-  virtual void addIndirectRoot(const Path & path) override;
-
+#if !NIX_IS_AT_LEAST(2,5,0)
   virtual void syncWithGC() override;
+#endif
 
+#if !NIX_IS_AT_LEAST(2,7,0)
   virtual Roots findRoots(bool censor) override;
 
   virtual void collectGarbage(const GCOptions & options, GCResults & results) override;
+
+  virtual void addIndirectRoot(const Path & path) override;
+#endif
 
   virtual void optimiseStore() override;
 
@@ -105,7 +135,12 @@ public:
       StorePathSet & willBuild, StorePathSet & willSubstitute, StorePathSet & unknown,
       uint64_t & downloadSize, uint64_t & narSize) override;
 
-  virtual std::shared_ptr<std::string> getBuildLog(const StorePath & path) override;
+#if NIX_IS_AT_LEAST(2,6,0)
+  virtual std::optional<std::string>
+#else
+  virtual std::shared_ptr<std::string>
+#endif
+    getBuildLog(const StorePath & path) override;
 
   virtual unsigned int getProtocol() override;
 
@@ -128,7 +163,12 @@ public:
 
   // Overrides
 
+#if NIX_IS_AT_LEAST(2,5,0)
+  virtual void queryRealisationUncached(const DrvOutput &,
+        Callback<std::shared_ptr<const Realisation>> callback) noexcept override;
+#else
   virtual std::optional<const Realisation> queryRealisation(const DrvOutput &) override;
+#endif
 
   virtual void ensurePath(const StorePath & path) override;