+++ hercules-ci-agent/hercules-ci-agent/Hercules/Agent/Compat.hs
@@ -5,8 +5,8 @@ import qualified Katip as K
 
 #if MIN_VERSION_katip(0,8,0)
 katipLevel :: K.Severity -> K.PermitFunc
-katipLevel =
-  K.permitItem
+katipLevel a =
+  K.permitItem a
 #else
 katipLevel x = x
 #endif