+++ hercules-ci-cnix-expr/src/Hercules/CNix/Expr.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleInstances #-}
@@ -43,6 +44,7 @@ import qualified Language.C.Inline.Cpp.Exceptions as C
 import Paths_hercules_ci_cnix_expr (getDataFileName)
 import Protolude hiding (evalState)
 import System.Directory (makeAbsolute)
+import Data.Aeson.KeyMap (toMapText)
 
 C.context (Hercules.CNix.Store.Context.context <> Hercules.CNix.Expr.Context.evalContext)
 
@@ -67,8 +69,6 @@ C.include "<nix/common-eval-args.hh>"
 C.include "<nix/get-drvs.hh>"
 
 C.include "<nix/derivations.hh>"
-
-C.include "<nix/affinity.hh>"
 
 C.include "<nix/globals.hh>"
 
@@ -241,7 +241,7 @@ isDerivation evalState (RawValue v) =
   (0 /=)
     <$> [C.throwBlock| int {
           if ($(Value *v) == NULL) { throw std::invalid_argument("forceValue value must be non-null"); }
-          $(EvalState *evalState)->forceValue(*$(Value *v));
+          $(EvalState *evalState)->forceValue(*$(Value *v), nix::noPos);
           return $(EvalState *evalState)->isDerivation(*$(Value *v));
         }|]
 
@@ -314,8 +314,12 @@ getDrvFile evalState (RawValue v) = liftIO do
       if (!drvInfo)
         throw EvalError("Not a valid derivation");
 
+#if NIX_IS_AT_LEAST(2,7,0)
+      StorePath storePath = drvInfo->requireDrvPath();
+#else
       std::string drvPath = drvInfo->queryDrvPath();
       StorePath storePath = state.store->parseStorePath(drvPath);
+#endif
 
       // write it (?)
       auto drv = state.store->derivationFromPath(storePath);
@@ -401,7 +405,7 @@ mkPath path =
             =<< [C.throwBlock| Value *{
       Value *r = new (NoGC) Value();
       std::string s($bs-ptr:path, $bs-len:path);
-      mkPath(*r, s.c_str());
+      r->mkPath(s.c_str());
       return r;
   }|]
         )
@@ -470,7 +474,7 @@ class ToRawValue a where
   toRawValue evalState a = rtValue <$> toValue evalState a
 
 class ToRawValue a => ToValue a where
-  type NixTypeFor a :: *
+  type NixTypeFor a :: Type
   toValue :: Ptr EvalState -> a -> IO (Value (NixTypeFor a))
 
 -- | Marshall values from Nix into Haskell. Instances must satisfy the
@@ -504,7 +508,7 @@ instance ToValue C.CBool where
     coerce
       <$> [C.block| Value *{
       Value *r = new (NoGC) Value();
-      mkBool(*r, $(bool b));
+      r->mkBool($(bool b));
       return r;
     }|]
 
@@ -525,7 +529,7 @@ instance ToValue Int64 where
     coerce
       <$> [C.block| Value *{
     Value *r = new (NoGC) Value();
-    mkInt(*r, $(int64_t i));
+    r->mkInt($(int64_t i));
     return r;
   }|]
 
@@ -543,7 +547,7 @@ instance ToValue C.CDouble where
     coerce
       <$> [C.block| Value *{
         Value *r = new (NoGC) Value();
-        mkFloat(*r, $(double f));
+        r->mkFloat($(double f));
         return r;
       }|]
 
@@ -557,11 +561,18 @@ instance ToValue Double where
 instance ToValue ByteString where
   type NixTypeFor ByteString = NixString
   toValue _ s =
+    -- TODO simplify when r->mkString(string_view) is safe in all supported Nix versions
     coerce
       <$> [C.block| Value *{
     Value *r = new (NoGC) Value();
-    std::string s($bs-ptr:s, $bs-len:s);
-    mkString(*r, s, {});
+    std::string_view s($bs-ptr:s, $bs-len:s);
+    // If empty, the pointer may be invalid; don't use it.
+    if (s.size() == 0) {
+      r->mkString("");
+    }
+    else {
+      r->mkString(GC_STRNDUP(s.data(), s.size()));
+    }
     return r;
   }|]
 
@@ -578,8 +589,45 @@ instance ToValue Text where
 
 instance ToRawValue a => ToRawValue (Map ByteString a)
 
+#if NIX_IS_AT_LEAST(2,6,0)
+withBindingsBuilder :: Integral n => Ptr EvalState -> n -> (Ptr BindingsBuilder' -> IO ()) -> IO (Value NixAttrs)
+withBindingsBuilder evalState n f = do
+  withBindingsBuilder' evalState n \bb -> do
+    f bb
+    v <- [C.block| Value* {
+      auto v = new (NoGC) Value();
+      v->mkAttrs(*$(BindingsBuilder *bb));
+      return v;
+    }|]
+    Value <$> mkRawValue v
+
+withBindingsBuilder' :: Integral n => Ptr EvalState -> n -> (Ptr BindingsBuilder' -> IO a) -> IO a
+withBindingsBuilder' evalState n =
+  let l :: C.CInt
+      l = fromIntegral n
+  in
+    bracket
+      [C.block| BindingsBuilder* {
+        auto &evalState = *$(EvalState *evalState);
+        return new BindingsBuilder(evalState, evalState.allocBindings($(int l)));
+      }|]
+      \bb -> [C.block| void { delete $(BindingsBuilder *bb); }|]
+#endif
+
 instance ToRawValue a => ToValue (Map ByteString a) where
   type NixTypeFor (Map ByteString a) = NixAttrs
+
+#if NIX_IS_AT_LEAST(2,6,0)
+  toValue evalState attrs = withBindingsBuilder evalState (length attrs) \bb -> do
+    attrs & traverseWithKey_ \k a -> do
+      RawValue aRaw <- toRawValue evalState a
+      [C.block| void {
+          EvalState &evalState = *$(EvalState *evalState);
+          std::string k($bs-ptr:k, $bs-len:k);
+          Value &a = *$(Value *aRaw);
+          $(BindingsBuilder *bb)->alloc(evalState.symbols.create(k)) = a;
+        }|]
+#else
   toValue evalState attrs = do
     let l :: C.CInt
         l = fromIntegral (length attrs)
@@ -602,6 +650,7 @@ instance ToRawValue a => ToValue (Map ByteString a) where
         $(Value *v)->attrs->sort();
       }|]
     Value <$> mkRawValue v
+#endif
 
 instance ToRawValue a => ToRawValue (Map Text a)
 
@@ -614,14 +663,14 @@ mkNull =
   coerce
     <$> [C.block| Value* {
           Value *v = new (NoGC) Value();
-          mkNull(*v);
+          v->mkNull();
           return v;
         }|]
 
 instance ToRawValue A.Value where
   toRawValue es (A.Bool b) = toRawValue es b
   toRawValue es (A.String s) = toRawValue es s
-  toRawValue es (A.Object fs) = toRawValue es fs
+  toRawValue es (A.Object fs) = toRawValue es $ toMapText fs
   toRawValue _es A.Null = mkNull
   toRawValue es (A.Number n) | Just i <- Sci.toBoundedInteger n = toRawValue es (i :: Int64)
   toRawValue es (A.Number f) = toRawValue es (Sci.toRealFloat f :: Double)
@@ -641,6 +690,19 @@ instance ToRawValue a => ToRawValue (H.HashMap Text a)
 
 instance ToRawValue a => ToValue (H.HashMap Text a) where
   type NixTypeFor (H.HashMap Text a) = NixAttrs
+
+#if NIX_IS_AT_LEAST(2,6,0)
+  toValue evalState attrs = withBindingsBuilder evalState (length attrs) \bb -> do
+    attrs & hmTraverseWithKey_ \k' a -> do
+      RawValue aRaw <- toRawValue evalState a
+      let k = encodeUtf8 k'
+      [C.block| void {
+          EvalState &evalState = *$(EvalState *evalState);
+          std::string k($bs-ptr:k, $bs-len:k);
+          Value &a = *$(Value *aRaw);
+          $(BindingsBuilder *bb)->alloc(evalState.symbols.create(k)) = a;
+        }|]
+#else
   toValue evalState attrs = do
     let l :: C.CInt
         l = fromIntegral (length attrs)
@@ -664,6 +726,7 @@ instance ToRawValue a => ToValue (H.HashMap Text a) where
         $(Value *v)->attrs->sort();
       }|]
     Value <$> mkRawValue v
+#endif
 
 instance ToRawValue a => ToRawValue (Vector a)