+++ hercules-ci-cnix-store/src/Hercules/CNix.hs
@@ -41,8 +41,6 @@ C.include "<nix/get-drvs.hh>"
 
 C.include "<nix/derivations.hh>"
 
-C.include "<nix/affinity.hh>"
-
 C.include "<nix/globals.hh>"
 
 C.include "hercules-ci-cnix/store.hxx"