+++ hercules-ci-cnix-expr/CHANGELOG.md
@@ -21,6 +21,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
  - Flakes are enabled during `init` on Nix 2.4pre
 
+## 0.2.0.2 - 2022-03-09
+
+### Added
+
+ - Nix 2.7 support
+
+## 0.2.0.1 - 2022-03-07
+
+### Fixed
+
+ - Build with newer Nix versions 2.5, 2.6
+
+### Added
+
+ - Improved conditional code support with `cabal-pkg-config-version-hook`
+
 ## 0.2.0.0 - 2021-06-22
 
 ### Added